+++ aim/web/ui/src/components/kit_v2/Link/Link.style.ts
@@ -1,11 +1,24 @@
 import { NavLink } from 'react-router-dom';
 
 import { styled, css } from 'config/stitches';
+import { textEllipsis } from 'config/stitches/foundations/layout';
 
 const StyledLink = css({
   textDecoration: 'none',
-  '&:hover': {
-    textDecoration: 'underline',
+  display: 'block',
+  variants: {
+    underline: {
+      true: {
+        '&:hover': {
+          textDecoration: 'underline',
+        },
+      },
+      false: {},
+    },
+    ellipsis: {
+      true: textEllipsis,
+      false: {},
+    },
   },
 });
 const StyledAnchor = styled('a', StyledLink);