+++ aim/web/ui/src/components/kit_v2/Select/Select.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FixedSizeList as List } from 'react-window';
 
-import { IconCaretRight, IconCaretUp } from '@tabler/icons-react';
+import { IconCaretDown, IconCaretUp } from '@tabler/icons-react';
 
 import Popover from '../Popover';
 import Button from '../Button';
@@ -22,6 +22,7 @@ const sizeDict = {
  * @description Virtualized Select component with search
  * @param {boolean} multiple - whether multiple select
  * @param {React.ReactNode} trigger - trigger element
+ * @param triggerProps
  * @param {PopoverProps} popoverProps - popover props
  * @param {string | string[] | undefined } value - selected value
  * @param {(val: string | string[]) => void} onValueChange - on value change callback
@@ -34,6 +35,7 @@ const sizeDict = {
 const Select = ({
   multiple,
   trigger,
+  triggerProps,
   popoverProps,
   value,
   onValueChange,
@@ -168,20 +170,23 @@ const Select = ({
 
   return (
     <Popover
-      {...popoverProps}
-      popperProps={{ css: { p: '$5 0' } }}
+      popperProps={{
+        ...popoverProps,
+        css: { p: '$5 0', ...popoverProps?.css },
+      }}
       trigger={({ open }) =>
-        trigger || multiple ? (
-          <Button size={size}>Select</Button>
-        ) : (
-          <Button
-            variant='outlined'
-            color='secondary'
-            rightIcon={open ? <IconCaretUp /> : <IconCaretRight />}
-          >
-            <Text>{triggerPlaceholder}</Text>
-          </Button>
-        )
+        typeof trigger === 'function'
+          ? trigger(open)
+          : trigger || (
+              <Button
+                variant='outlined'
+                color='secondary'
+                rightIcon={open ? <IconCaretUp /> : <IconCaretDown />}
+                {...triggerProps}
+              >
+                <Text css={{ flex: '1' }}>{triggerPlaceholder}</Text>
+              </Button>
+            )
       }
       content={
         <>