+++ aim/web/ui/src/components/kit_v2/Text/Text.d.ts
@@ -11,6 +11,9 @@ export interface ITextProps
   disabled?: boolean;
   size?: CSS['fontSize'];
   color?: CSS['color'];
+  ellipsis?: boolean;
+  textTransform?: CSS['textTransform'];
+  lineHeight?: CSS['lineHeight'];
   css?: CSS;
 }