+++ aim/web/ui/src/pages/Dashboard/components/ProjectStatistics/ProjectStatistics.d.ts
@@ -9,4 +9,5 @@ export interface IProjectStatistic {
   icon?: IconName;
   navLink?: string;
   badge?: { value: string; style?: React.CSSProperties };
+  description?: string;
 }